# apps/server/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BridgeAI API",
    "version": "1.0.0",
    "description": "BridgeAI Backend API - AI-powered matching and communication platform.\n\n## API Versioning\n\nAll endpoints are versioned under `/api/v1/`. When breaking changes are introduced, a new version prefix (e.g., `/api/v2/`) will be created.\n\n**Current version**: v1\n\n**Versioning strategy**:\n- URL path versioning (`/api/v1/`, `/api/v2/`)\n- Backward-compatible changes (new fields, new endpoints) do not bump version\n- Breaking changes (removed fields, changed types) bump version\n- Old versions remain available for at least 6 months after deprecation\n\n## Authentication\n\nAll authenticated endpoints require a Bearer token in the Authorization header:\n`Authorization: Bearer <token>`\n\nTokens are obtained via `POST /api/v1/auth/login` or `POST /api/v1/auth/register`.\n\n## Rate Limiting\n\nAPI requests are rate-limited by IP address. Rate limit headers are included in responses:\n- `X-RateLimit-Limit`: Maximum requests per window\n- `X-RateLimit-Remaining`: Remaining requests in current window\n- `X-RateLimit-Reset`: Timestamp when the rate limit resets\n\n## Response Format\n\nAll responses follow a standard format:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"message\": \"Optional message\"\n}\n```\n\nError responses:\n```json\n{\n  \"success\": false,\n  \"error\": \"Error message\",\n  \"code\": \"ERROR_CODE\"\n}\n```",
    "contact": {
...
    },
...
  },
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token. Obtain via POST /api/v1/auth/login"
      }
    },
...
  },
...
  "paths": {
...
    "/admin/security/stats": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get security stats",
        "description": "Security overview dashboard statistics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Security stats"
          }
        }
      }
    },
    "/admin/security/realtime": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get real-time security data",
        "description": "Real-time security monitoring data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Real-time data"
          }
        }
      }
    },
    "/admin/security/ddos": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get DDoS stats",
        "description": "DDoS protection statistics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "DDoS stats"
          }
        }
      }
    },
    "/admin/security/blocked-ips": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get blocked IPs",
        "description": "List of blocked IP addresses",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Blocked IPs list"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Block IP",
        "description": "Add an IP to the block list",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ip": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "IP blocked"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Unblock IP",
        "description": "Remove an IP from the block list",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ip",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "IP unblocked"
          }
        }
      }
    },
    "/admin/security/whitelist": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get whitelisted IPs",
        "description": "List of whitelisted IP addresses",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Whitelist"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Add to whitelist",
        "description": "Add an IP to the whitelist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ip": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added to whitelist"
          }
        }
      }
    },
    "/admin/security/events": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get security events",
        "description": "List security events with filtering",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Security events"
          }
        }
      }
    },
    "/admin/security/events/{id}/resolve": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Resolve security event",
        "description": "Mark a security event as resolved",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resolution": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved"
          }
        }
      }
    },
...
  }
}